requirements.txt
APScheduler==3.6.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
decorator==5.1.1
idna==3.6
pycparser==2.21
python-dotenv==0.9.1
python-telegram-bot==13.0
pytz==2024.1
requests==2.31.0
requests-toolbelt==0.10.1
setuptools==69.2.0
six==1.16.0
tornado==6.4
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.15